refactor(web): platformとlibの責務を分割
