+++ apps/web/components.json
@@ -12,7 +12,7 @@
   },
   "aliases": {
     "components": "@/components",
-    "utils": "@/lib/utils",
+    "utils": "@/lib/class-names",
     "ui": "@/components/ui",
     "lib": "@/lib",
     "hooks": "@/hooks"
