+++ apps/web/src/features/blog/components/BlogSkeleton.tsx
@@ -1,4 +1,4 @@
-import { cn } from '@/lib/utils';
+import { cn } from '@/lib/class-names';
 import { Skeleton } from '@dayopt/components';
 
 interface BlogSkeletonProps {
