+++ apps/web/src/features/blog/lib/blog.ts
@@ -3,8 +3,13 @@ import {
   blogFrontMatterSchema,
   parseFrontMatter,
 } from '@/lib/content-schemas';
-import { createStructuredError, ErrorCategory, ErrorLevel, logError } from '@/lib/error-utils';
-import { calculateReadingTime } from '@/lib/utils';
+import {
+  createStructuredError,
+  ErrorCategory,
+  ErrorLevel,
+  logError,
+} from '@/lib/errors/structured-error';
+import { calculateReadingTime } from '@/lib/reading-time';
 import fs from 'fs';
 import matter from 'gray-matter';
 import path from 'path';
