+++ apps/web/src/platform/config/__tests__/env.test.ts
@@ -1,6 +1,6 @@
-import { describe, expect, it } from 'vitest';
+import { afterEach, describe, expect, it, vi } from 'vitest';
 
-import { assertProductionRuntimeEnv, type EnvConfig } from '../env';
+import { assertProductionRuntimeEnv, EnvValidationError, loadEnv, type EnvConfig } from '../env';
 
 function validProductionRuntimeEnv(): Partial<EnvConfig> {
   return {
@@ -18,41 +18,104 @@ function validProductionRuntimeEnv(): Partial<EnvConfig> {
   };
 }
 
+afterEach(() => {
+  vi.restoreAllMocks();
+  vi.unstubAllEnvs();
+});
+
 describe('assertProductionRuntimeEnv', () => {
-  it('accepts runtime Sentry DSNs without build-only upload credentials', () => {
+  it('完全な Vercel Production env を受け入れる', () => {
     expect(() => assertProductionRuntimeEnv(validProductionRuntimeEnv())).not.toThrow();
   });
 
-  it('requires both browser and server DSNs in Vercel Production', () => {
-    const missingBrowserDsn = validProductionRuntimeEnv();
-    delete missingBrowserDsn.NEXT_PUBLIC_SENTRY_DSN;
-
-    expect(() => assertProductionRuntimeEnv(missingBrowserDsn)).toThrow(
-      'NEXT_PUBLIC_SENTRY_DSN is required for Web Sentry in Vercel Production',
-    );
-  });
-
-  it('requires both Turnstile credentials in Vercel Production', () => {
-    const missingTurnstileSecret = validProductionRuntimeEnv();
-    delete missingTurnstileSecret.TURNSTILE_SECRET_KEY;
-
-    expect(() => assertProductionRuntimeEnv(missingTurnstileSecret)).toThrow(
-      'NEXT_PUBLIC_TURNSTILE_SITE_KEY and TURNSTILE_SECRET_KEY are required in production',
-    );
-  });
-
-  it('requires an explicit contact repository in Vercel Production', () => {
-    const missingContactRepository = validProductionRuntimeEnv();
-    delete missingContactRepository.GITHUB_CONTACT_REPO;
-
-    expect(() => assertProductionRuntimeEnv(missingContactRepository)).toThrow(
-      'GITHUB_CONTACT_REPO is required for the private contact repository',
-    );
-  });
-
-  it('does not require Sentry runtime variables in Preview', () => {
+  it('CI と Preview では production secret を要求しない', () => {
+    expect(() =>
+      assertProductionRuntimeEnv({ NODE_ENV: 'production', VERCEL_ENV: 'production', CI: true }),
+    ).not.toThrow();
     expect(() =>
       assertProductionRuntimeEnv({ NODE_ENV: 'production', VERCEL_ENV: 'preview' }),
     ).not.toThrow();
+  });
+
+  it('Production の不足項目を既存順序でまとめて報告する', () => {
+    let thrown: unknown;
+    try {
+      assertProductionRuntimeEnv({ NODE_ENV: 'production', VERCEL_ENV: 'production' });
+    } catch (error) {
+      thrown = error;
+    }
+
+    expect(thrown).toBeInstanceOf(EnvValidationError);
+    const message = (thrown as Error).message;
+    const expectedOrder = [
+      'NEXT_PUBLIC_APP_URL or VERCEL_URL',
+      'GITHUB_TOKEN is not set',
+      'GITHUB_CONTACT_REPO is required',
+      'NEXT_PUBLIC_SENTRY_DSN is required',
+      'SENTRY_DSN is required',
+      'NEXT_PUBLIC_TURNSTILE_SITE_KEY and TURNSTILE_SECRET_KEY',
+      'UPSTASH_REDIS_REST_URL and UPSTASH_REDIS_REST_TOKEN are required',
+    ];
+
+    let previousIndex = -1;
+    for (const part of expectedOrder) {
+      const index = message.indexOf(part);
+      expect(index).toBeGreaterThan(previousIndex);
+      previousIndex = index;
+    }
+  });
+
+  it('Upstash の片側だけがある場合は pair と required の両方を報告する', () => {
+    const env = validProductionRuntimeEnv();
+    delete env.UPSTASH_REDIS_REST_TOKEN;
+
+    expect(() => assertProductionRuntimeEnv(env)).toThrow(
+      'Both UPSTASH_REDIS_REST_URL and UPSTASH_REDIS_REST_TOKEN should be set together',
+    );
+    expect(() => assertProductionRuntimeEnv(env)).toThrow(
+      'UPSTASH_REDIS_REST_URL and UPSTASH_REDIS_REST_TOKEN are required in production',
+    );
+  });
+
+  it('引数省略時は現在の process.env を読み直す', () => {
+    vi.stubEnv('NODE_ENV', 'production');
+    vi.stubEnv('VERCEL_ENV', 'preview');
+    expect(() => assertProductionRuntimeEnv()).not.toThrow();
+  });
+});
+
+describe('loadEnv', () => {
+  it('boolean と privacy mode を既存規則で parse する', () => {
+    vi.stubEnv('NODE_ENV', 'test');
+    vi.stubEnv('CI', '0');
+    vi.stubEnv('PRIVACY_PROTECTION_MODE', 'strict');
+
+    expect(loadEnv()).toMatchObject({
+      NODE_ENV: 'test',
+      CI: false,
+      PRIVACY_PROTECTION_MODE: 'strict',
+    });
+  });
+
+  it('development で不足・片側設定を3件 warning する', () => {
+    vi.stubEnv('NODE_ENV', 'development');
+    vi.stubEnv('GITHUB_TOKEN', '');
+    vi.stubEnv('NEXT_PUBLIC_TURNSTILE_SITE_KEY', 'configured');
+    vi.stubEnv('TURNSTILE_SECRET_KEY', '');
+    vi.stubEnv('UPSTASH_REDIS_REST_URL', 'https://example.upstash.io');
+    vi.stubEnv('UPSTASH_REDIS_REST_TOKEN', '');
+    const warn = vi.spyOn(console, 'warn').mockImplementation(() => undefined);
+
+    loadEnv();
+
+    expect(warn.mock.calls).toEqual([
+      ['[ENV WARNING] GITHUB_TOKEN is not set. Contact form will not work.'],
+      [
+        '[ENV WARNING] Both NEXT_PUBLIC_TURNSTILE_SITE_KEY and TURNSTILE_SECRET_KEY should be set together.',
+      ],
+      [
+        '[ENV WARNING] Both UPSTASH_REDIS_REST_URL and UPSTASH_REDIS_REST_TOKEN should be set together.',
+      ],
+    ]);
   });
 });
